% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.bison.R
\name{as.bison}
\alias{as.bison}
\alias{as.bison.bisonkey}
\alias{as.bison.character}
\alias{as.bison.data.frame}
\alias{as.bison.list}
\alias{as.bison.numeric}
\alias{as.bison.occdat}
\title{Coerce occurrence keys to bisonkey/occkey objects}
\usage{
as.bison(x)

\method{as.bison}{bisonkey}(x)

\method{as.bison}{bisonkey}(x)

\method{as.bison}{occdat}(x)

\method{as.bison}{data.frame}(x)

\method{as.bison}{numeric}(x)

\method{as.bison}{character}(x)

\method{as.bison}{list}(x)
}
\arguments{
\item{x}{Various inputs, including the output from a call to \code{\link{occ}} (class occdat),
\code{\link{occ2df}} (class data.frame), or a list, numeric, character, or gbifkey, or occkey.}
}
\value{
One or more in a list of both class gbifkey and occkey
}
\description{
Coerce occurrence keys to bisonkey/occkey objects
}
\details{
Internally, we use \code{\link[rgbif]{occ_get}}, whereas \code{\link{occ}}
uses \code{\link[rgbif]{occ_search}}. We can use \code{\link[rgbif]{occ_get}} here
because we have the occurrence key to go directly to the occurrence record.
}
\examples{
\dontrun{
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Carduelis tristis')
out <- occ(query=spnames, from='bison', limit=2)
res <- occ2df(out)
(tt <- as.bison(out))
(uu <- as.bison(res))
as.bison(as.numeric(res$key[1]))
as.bison(res$key[1])
as.bison(as.list(res$key[1:2]))
as.bison(tt[[1]])
as.bison(uu[[1]])
as.bison(tt[1:2])
}
}

