% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nbn.R
\name{as.nbn}
\alias{as.nbn}
\title{Coerce occurrence keys to NBN id objects}
\usage{
as.nbn(x)
}
\arguments{
\item{x}{Various inputs, including the output from a call to \code{\link{occ}}
(class occdat), \code{\link{occ2df}} (class data.frame), or a list, numeric,
character, nbnkey, or occkey.}
}
\value{
One or more in a list of both class nbnkey and occkey
}
\description{
Coerce occurrence keys to NBN id objects
}
\examples{
\dontrun{
spnames <- c('Mola mola', 'Loligo vulgaris', 'Stomias boa')
out <- occ(query=spnames, from='nbn', limit=2)
res <- occ2df(out)
(tt <- as.nbn(out))
(uu <- as.nbn(res))
as.nbn(x = res$key[1])
as.nbn(as.list(res$key[1:2]))
as.nbn(tt[[1]])
as.nbn(uu[[1]])
as.nbn(tt[1:2])

library("dplyr")
rbind_all(lapply(tt, function(x) data.frame(unclass(x)$data)))
}
}
\keyword{internal}

