% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_spocc.r
\name{clean_spocc}
\alias{clean_spocc}
\title{Clean spocc data}
\usage{
clean_spocc(input)
}
\arguments{
\item{input}{An object of class occdat}
}
\value{
Returns an object of class occdat+occlean. See attributes of the return object for
details on cleaning results.
}
\description{
Clean spocc data
}
\details{
We'll continue to add options for cleaning data, but for now, this function:

\itemize{
 \item Removes impossible values of latitude and longitude
 \item Removes any NA values of latitude and longitude
 \item Removes points at 0,0 - these points are likely wrong
}
}
\examples{
\dontrun{
res <- occ(query = c('Ursus','Accipiter','Rubus'), from = 'bison', limit=10)
class(res)
res_cleaned <- clean_spocc(input=res)
class(res_cleaned) # now with classes occdat and occclean
}
}

