% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixnames.r
\name{fixnames}
\alias{fixnames}
\title{Change names to be the same for each taxon.}
\usage{
fixnames(...)
}
\arguments{
\item{...}{ignored}
}
\description{
DEFUNT: This function has moved to \code{scrubr::fix_names}
}
\keyword{internal}
