\name{get_colors}
\alias{get_colors}
\title{Get colors from a vector of input taxonomic names, and palette}
\usage{
get_colors(vec, palette_name)
}
\arguments{
  \item{vec}{Vector of strings}

  \item{palette_name}{Palette name}
}
\description{
Get colors from a vector of input taxonomic names, and
palette
}
\keyword{internal}

