% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_rcharts.r
\name{get_palette}
\alias{get_palette}
\title{Get palette actual name from longer names}
\usage{
get_palette(userselect)
}
\arguments{
\item{userselect}{User input}
}
\description{
Get palette actual name from longer names
}

