% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis_helpers.R
\name{obis_search}
\alias{obis_search}
\title{OBIS search}
\usage{
obis_search(scientificName = NULL, size = 500, after = NULL,
  taxonid = NULL, aphiaid = NULL, areaid = NULL, datasetid = NULL,
  instituteid = NULL, nodeid = NULL, startdate = NULL,
  enddate = NULL, startdepth = NULL, enddepth = NULL,
  geometry = NULL, exclude = NULL, fields = NULL, ...)
}
\arguments{
\item{size}{(integer) number of results to fetch}

\item{after}{(character) Occurrence UUID up to which to skip.}

\item{taxonid}{(character) Taxon AphiaID.}

\item{areaid}{(character) Area ID.}

\item{datasetid}{(character) Dataset UUID.}

\item{instituteid}{(character) Institute ID.}

\item{nodeid}{(character) Node UUID.}

\item{startdate}{(character) Start date formatted as YYYY-MM-DD.}

\item{enddate}{(character) End date formatted as YYYY-MM-DD.}

\item{startdepth}{(integer) Start depth, in meters.}

\item{enddepth}{(integer) End depth, in meters.}

\item{geometry}{(character) Geometry, formatted as WKT.}

\item{exclude}{(character) set of quality flags to be excluded.
one or more in a vector}

\item{fields}{(character) Field to be included in the result set.
one or more in a vector}

\item{scientificname}{(character) Scientific name. Leave empty to
include all taxa. This is what we pass your name query to}
}
\description{
OBIS search
}
\keyword{internal}
