% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occ2df.R
\name{occ2df}
\alias{occ2df}
\title{Combine results from occ calls to a single data.frame}
\usage{
occ2df(obj, what = "data")
}
\arguments{
\item{obj}{Input from occ}

\item{what}{(character) One of data (default) or all (with metadata)}
}
\description{
Combine results from occ calls to a single data.frame
}
\details{
This function combines a subset of data from each data provider to a single data.frame, or
metadata plus data if you request \code{what="all"}. The single data.frame contains the
following columns:

\itemize{
 \item name - scientific (or common) name
 \item longitude - decimal degree longitude
 \item latitude - decimal degree latitude
 \item prov - data provider
 \item date - occurrence record date
 \item key - occurrence record key
}

AntWeb doesn't provide dates, so occurrence rows from that provider are blank.

Ecoengine doesn't provide unique keys for each occurrence record, so we generate one
for each record. We cache records returned from the \code{\link{occ}} function on your
machine, and you can use the keys from the output of this function to request
further output in \code{inspect}, which uses that cached data. Other data sources
provide unique keys already, so we can query the database online again for more data.
}
\examples{
\dontrun{
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Carduelis tristis')
out <- occ(query=spnames, from='gbif', gbifopts=list(hasCoordinate=TRUE), limit=10)
occ2df(out)
}
}

