% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_coverage.R
\name{occ_coverage}
\alias{occ_coverage}
\title{Automatically generate coverages for a spocc search}
\usage{
occ_coverage(occObj, coverage = "all")
}
\arguments{
\item{occObj}{an search object returned by occ}

\item{coverage}{a vector of coverage types to generate.  These include 'temporal','spatial','taxa', or just 'all'.}
}
\description{
This function will automatically generate metadata for spocc queries that can then be converted to other standards.
}
\keyword{internal}

