\name{occ_todf}
\alias{occ_todf}
\title{Coerce elements of output from a single occ() call to a single data.frame}
\usage{
  occ_todf(x)
}
\arguments{
  \item{x}{An object of class occdat}
}
\value{
  A data.frame
}
\description{
  Coerce elements of output from a single occ() call to a
  single data.frame
}

