\name{occDatSpp-class}
\alias{occDatSpp}
\alias{occDatSpp-class}
\title{"occDatSpp" class}
\description{
  "occDatSpp" class
}
\seealso{
  Other res: \code{\link{occDat-class}},
  \code{\link{occDat-class}}, \code{\link{occDf-class}},
  \code{\link{occDfMany-class}},
  \code{\link{occResult-class}},
  \code{\link{occResultList-class}},
  \code{\link{occdf-class}}, \code{\link{occdfmany-class}}
}

