\name{occdf-class}
\alias{occdf-class}
\title{Coerce to sp object}
\description{
  Coerce to sp object
}
\seealso{
  Other res: \code{\link{occDat-class}},
  \code{\link{occDat-class}}, \code{\link{occDf-class}},
  \code{\link{occResult-class}},
  \code{\link{occdfmany-class}},
  \code{\link{occdfmany-class}}
}

