\name{occResult-class}
\alias{occResult}
\alias{occResult-class}
\title{"occResult" class}
\description{
  "occResult" class
}
\seealso{
  Other res: \code{\link{occDat-class}},
  \code{\link{occDat-class}}, \code{\link{occDf-class}},
  \code{\link{occdf-class}}, \code{\link{occdfmany-class}},
  \code{\link{occdfmany-class}}
}

