\name{occResultList-class}
\alias{occResultList}
\alias{occResultList-class}
\title{"occResultList" class}
\description{
  "occResultList" class
}
\seealso{
  Other res: \code{\link{occDat-class}},
  \code{\link{occDat-class}},
  \code{\link{occDatSpp-class}}, \code{\link{occDf-class}},
  \code{\link{occDfMany-class}},
  \code{\link{occResult-class}}, \code{\link{occdf-class}},
  \code{\link{occdfmany-class}}
}

