% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.occdat}
\alias{print.occdat}
\title{Print brief summary of occ function output}
\usage{
\method{print}{occdat}(x, ...)
}
\arguments{
\item{x}{Input...}

\item{...}{Ignored.}
}
\description{
Print brief summary of occ function output
}
\examples{
\dontrun{
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Spinus tristis')
out <- occ(query = spnames, from = 'gbif', gbifopts = list(hasCoordinate=TRUE))
print(out)
out # gives the same thing

# you can still drill down into the data easily
out$gbif$meta
out$gbif$data
}
}

