% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.occdatind}
\alias{print.occdatind}
\title{Print method for individual data sources}
\usage{
\method{print}{occdatind}(x, ..., n = 10)
}
\arguments{
\item{x}{Input, of class occdatind}

\item{...}{Further args, ignored}

\item{n}{Number of data frame rows to print}
}
\description{
Print method for individual data sources
}

