% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spocc-package.R
\docType{package}
\name{spocc-package}
\alias{spocc}
\alias{spocc-package}
\title{spocc: Interface to Species Occurrence Data Sources}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A programmatic interface to many species occurrence data sources, including Global Biodiversity Information Facility ('GBIF'), 'iNaturalist', 'eBird', Integrated Digitized 'Biocollections' ('iDigBio'), 'VertNet', Ocean 'Biogeographic' Information System ('OBIS'), and Atlas of Living Australia ('ALA'). Includes functionality for retrieving species occurrence data, and combining those data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/spocc (devel)}
  \item \url{https://docs.ropensci.org/spocc/ (user manual)}
  \item Report bugs at \url{https://github.com/ropensci/spocc/issues/}
}

}
\author{
\strong{Maintainer}: Hannah Owens \email{hannah.owens@gmail.com} (\href{https://orcid.org/0000-0003-0071-1745}{ORCID})

Authors:
\itemize{
  \item Vijay Barve (\href{https://orcid.org/0000-0002-4852-2567}{ORCID})
  \item Scott Chamberlain (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})
}

Other contributors:
\itemize{
  \item Karthik Ram [contributor]
  \item Ted Hart [contributor]
  \item rOpenSci (ropensci.org/) [funder]
}

}
\keyword{internal}
