% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spocc_colClasses}
\alias{spocc_colClasses}
\title{Coerces data.frame columns to the specified classes}
\usage{
spocc_colClasses(d, colClasses)
}
\arguments{
\item{d}{A data.frame.}

\item{colClasses}{A vector of column attributes, one of:
numeric, factor, character, etc.}
}
\description{
Coerces data.frame columns to the specified classes
}
\examples{
\dontrun{
dat <- data.frame(xvar = seq(1:10), yvar = rep(c('a','b'),5)) # make a data.frame
str(dat)
str(colClasses(dat, c('factor','factor')))
}
}
\keyword{internal}

