% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_github.r
\name{spocc_create_gist}
\alias{spocc_create_gist}
\title{Function that takes a list of files and creates payload for API}
\usage{
spocc_create_gist(filenames, description = "", public = TRUE)
}
\arguments{
\item{filenames}{names of files to post}

\item{description}{brief description of gist (optional)}

\item{public}{whether gist is public (defaults to TRUE)}
}
\description{
Function that takes a list of files and creates payload for API
}
\keyword{internal}

