\name{spocc_parseresults}
\alias{spocc_parseresults}
\title{Parse results from call to occurrencelist endpoint}
\usage{
spocc_parseresults(x, ..., removeZeros = removeZeros)
}
\arguments{
  \item{x}{A list}

  \item{...}{Further args passed on to gbifxmlToDataFrame}

  \item{removeZeros}{remove zeros or not}
}
\description{
Parse results from call to occurrencelist endpoint
}
\keyword{internal}

