\name{spocc_toGeoJSON}
\alias{spocc_toGeoJSON}
\title{Get palette actual name from longer names}
\usage{
  spocc_toGeoJSON(list_, lat = "latitude",
    lon = "longitude")
}
\arguments{
  \item{list_}{A list}

  \item{lat}{Latitude name}

  \item{lon}{Longitude name}
}
\description{
  Get palette actual name from longer names
}
\keyword{internal}

