% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.r
\name{spocc_type_sum}
\alias{spocc_type_sum}
\alias{spocc_type_sum.Date}
\alias{spocc_type_sum.POSIXt}
\alias{spocc_type_sum.array}
\alias{spocc_type_sum.character}
\alias{spocc_type_sum.default}
\alias{spocc_type_sum.factor}
\alias{spocc_type_sum.integer}
\alias{spocc_type_sum.logical}
\alias{spocc_type_sum.matrix}
\alias{spocc_type_sum.numeric}
\title{Type summary}
\usage{
spocc_type_sum(x)

\method{spocc_type_sum}{default}(x)

\method{spocc_type_sum}{character}(x)

\method{spocc_type_sum}{Date}(x)

\method{spocc_type_sum}{factor}(x)

\method{spocc_type_sum}{integer}(x)

\method{spocc_type_sum}{logical}(x)

\method{spocc_type_sum}{array}(x)

\method{spocc_type_sum}{matrix}(x)

\method{spocc_type_sum}{numeric}(x)

\method{spocc_type_sum}{POSIXt}(x)
}
\description{
Type summary
}
\keyword{internal}

