\name{toGeoJSON}
\alias{toGeoJSON}
\title{Get palette actual name from longer names}
\usage{
  toGeoJSON(list_, lat = "latitude", lon = "longitude")
}
\arguments{
  \item{list_}{A list}

  \item{lat}{Latitude name}

  \item{lon}{Longitude name}
}
\description{
  Get palette actual name from longer names
}
\keyword{internal}

