% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_wkt.R
\name{wkt}
\alias{read_wkt}
\alias{wkt}
\alias{write_wkt}
\title{Read or write WKT}
\usage{
read_wkt(wkt)

write_wkt(geojson)
}
\arguments{
\item{wkt}{(character) A Well Known Text string}

\item{geojson}{(character) GeoJSON as a list, character string or JSON}
}
\description{
Read or write WKT
}
\examples{
wkt <- 'LINESTRING (30 10, 10 30, 40 40)'
read_wkt(wkt)

wkt <- "POLYGON((38.4 -125,40.9 -125,40.9 -121.8,38.4 -121.8,38.4 -125))"
read_wkt(wkt)
}

