% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ecoengine.R
\name{as.ecoengine}
\alias{as.ecoengine}
\title{Coerce occurrence keys to ecoenginekey/occkey objects}
\usage{
as.ecoengine(...)
}
\arguments{
\item{...}{ignored}
}
\description{
DEFUNCT
}
\keyword{internal}
