% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.idigbio.R
\name{as.idigbio}
\alias{as.idigbio}
\title{Coerce occurrence keys to idigbio uuid objects}
\usage{
as.idigbio(x)
}
\arguments{
\item{x}{Various inputs, including the output from a call to \code{\link{occ}}
(class occdat), \code{\link{occ2df}} (class data.frame), or a list, numeric,
character, gbifkey, or occkey.}
}
\value{
One or more in a list of both class idigbiokey and occkey
}
\description{
Coerce occurrence keys to idigbio uuid objects
}
\details{
Internally, we use \code{idig_search_records}.
}
\examples{
\dontrun{
# idig_search(rq=list(uuid="0b657c62-fa7b-400e-8f47-d05bb3bbc463"))
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Carduelis tristis')
out <- occ(query=spnames, from='idigbio', limit=2)
res <- occ2df(out)
(tt <- as.idigbio(out))
(uu <- as.idigbio(res))
as.idigbio(res$key[1])
as.idigbio(as.list(res$key[1:2]))
as.idigbio(tt[[1]])
as.idigbio(uu[[1]])
as.idigbio(tt[1:2])

library("dplyr")
rbind_all(lapply(tt, function(x) data.frame(unclass(x))))
}
}

