% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clean_spocc}
\alias{clean_spocc}
\title{Clean spocc data}
\usage{
clean_spocc(x, country = NULL, habitat = NULL)
}
\description{
Clean spocc data
}
\examples{
library(spocc)
res <- occ(query = 'Accipiter striatus', from = 'gbif', limit=60)
clean_spocc(x=res$gbif$data[[1]])
}

