% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clean_spocc}
\alias{clean_spocc}
\title{Clean spocc data}
\usage{
clean_spocc(input, country = NULL, habitat = NULL)
}
\arguments{
\item{input}{An object of class occdat}

\item{country}{(logical) Attempt to clean based on country}

\item{habitat}{(logical) Attempt to clean based on habitat}
}
\description{
Clean spocc data
}
\examples{
\dontrun{
res <- occ(query = c('Ursus','Accipiter','Rubus'), from = 'bison', limit=120)
res_cleaned <- clean_spocc(res)
class(res_cleaned) # now with classes occdat and occclean
}
}

