% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\alias{inspect.data.frame}
\alias{inspect.occdat}
\alias{inspect.occkey}
\title{Get more data on individual occurrences.}
\usage{
inspect(x, from = "gbif")

\method{inspect}{data.frame}(x, from = "gbif")

\method{inspect}{occdat}(x, from = "gbif")

\method{inspect}{occkey}(x, from = "gbif")
}
\arguments{
\item{x}{The output from \code{\link{occ}} call.}

\item{from}{The data provider. One of gbif, bison, inat, ebird, or ecoengine}
}
\description{
Get more data on individual occurrences.
}
\examples{
\dontrun{
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Carduelis tristis')
out <- occ(query=spnames, from=c('gbif','bison'), gbifopts=list(hasCoordinate=TRUE), limit=2)
res <- occ2df(out)
inspect(res)

out <- occ(query=spnames, from='gbif', gbifopts=list(hasCoordinate=TRUE), limit=4)
res <- occ2df(out)
inspect(res)

# from occkeys
key <- as.gbif(res$key[1])
inspect(key)
}
}

