% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Get more data on individual occurrences.}
\usage{
inspect(obj = NULL, keys = NULL, from = "gbif")
}
\arguments{
\item{obj}{The output from \code{\link{occ}} call.}

\item{keys}{You can alternatively pass in keys, in which case you have to pass a named vector
or the from param will be used to pick which source to query.}

\item{from}{The data provider}
}
\description{
Get more data on individual occurrences.
}
\details{
This function xxxx
}
\examples{
\dontrun{
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Carduelis tristis')
out <- occ(query=spnames, from=c('gbif','ebird'), gbifopts=list(hasCoordinate=TRUE), limit=4)
res <- occ2df(out)
inspect(res, from="gbif")
}
}

