% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mapggplot}
\alias{mapggplot}
\title{ggplot2 visualization of species occurences}
\usage{
mapggplot(df, zoom = 5, point_color = "#86161f")
}
\arguments{
\item{df}{Input \code{data.frame}}

\item{zoom}{The zoom level of the map that is passed on to \code{get_map}. The spocc default is 5 but values can range from 3 (for a continent level zoom) to 21 (zoomed to the closest bulding). Adjust the values within this range till the map looks to your satisfaction.}

\item{point_color}{Default color of your points}
}
\description{
ggplot2 visualization of species occurences
}
\examples{
\dontrun{
ecoengine_data <- occ(query = 'Lynx rufus californicus', from = 'ecoengine')
mapggplot(ecoengine_data)
gbif_data <- occ(query = 'Accipiter striatus', from = 'gbif')
mapggplot(gbif_data)
bison_data <- occ(query = 'Accipiter striatus', from = 'bison', bisonopts=list(count=100))
mapggplot(bison_data)
}
}

