% Generated by roxygen2 (4.0.1): do not edit by hand
\name{occ2df}
\alias{occ2df}
\title{Combine results from occ calls to a single data.frame}
\usage{
occ2df(obj, what = "data")
}
\arguments{
\item{obj}{Input from occ}

\item{what}{One of data (default) or all (with metadata)}
}
\description{
Combine results from occ calls to a single data.frame
}
\examples{
\dontrun{
spnames <- c('Accipiter striatus', 'Setophaga caerulescens', 'Spinus tristis')
out <- occ(query=spnames, from='gbif', gbifopts=list(hasCoordinate=TRUE))
occ2df(out)
}
}

