% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_options.r
\name{occ_options}
\alias{occ_options}
\title{Look up options for parameters passed to each source}
\usage{
occ_options(from = "gbif", where = "console")
}
\arguments{
\item{from}{(character) Data source to get data from, any combination of
gbif, bison, ebird, idigibio and/or vertnet. Case doesn't matter.
inat is not included here, see that package's help docs.}

\item{where}{(character) One of console (print to console) or html (opens
help page, if in non-interactive R session, prints help to console).}
}
\value{
Opens up the documentation for the function that is used internally
within the occ function for each source.
}
\description{
Look up options for parameters passed to each source
}
\details{
Any of the parameters passed to e.g. \code{\link[rgbif:occ_data]{rgbif::occ_data()}} from the
\code{rgbif} package can be passed in the associated gbifopts list
in \code{\link[=occ]{occ()}}

Note that the from parameter is lowercased within the function and is
called through match.arg first, so you can match on unique partial
strings too (e.g., 'rv' for 'rvertnet').

For some data sources we don't import the canonical package, but instead
have our own intenral helper functions and the package in question is
not imported, but we do suggest seeing the help for the package on
their parameters:

\itemize{
\item \strong{AntWeb}: \code{?AntWeb::aw_data}
\item \strong{ecoengine}: \code{?ecoengine::ee_observations}
}

For others,
}
\examples{
\dontrun{
# opens up documentation for this function
occ_options()

# Open up documentation for the appropriate search function for each source
occ_options('gbif')
occ_options('ebird')
occ_options('bison')
occ_options('idigbio')
occ_options('vertnet')

# Or open in html version
occ_options('bison', 'html')
}
}
