% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{spocc-package}
\alias{spocc}
\alias{spocc-package}
\title{R interface to many species occurrence data sources}
\description{
You don't need API keys for any of the data providers thus far. However, some
may be added in the future that require authentication, just FYI.
}
\details{
Currently supported species occurrence databases:

\tabular{ll}{
Provider \tab Web \cr
GBIF \tab \url{http://www.gbif.org/} \cr
BISON \tab \url{http://bison.usgs.ornl.gov/} \cr
iNaturalist \tab \url{http://www.inaturalist.org/} \cr
Berkeley ecoengine \tab \url{https://ecoengine.berkeley.edu/} \cr
AntWeb \tab \url{http://www.antweb.org/} \cr
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Karthik Ram \email{karthik.ram@gmail.com}

Ted Hart \email{edmund.m.hart@gmail.com}
}
\keyword{package}

