% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spocc-package.R
\docType{package}
\name{spocc-package}
\alias{spocc}
\alias{spocc-package}
\title{R interface to many species occurrence data sources}
\description{
Currently supported species occurrence databases:
}
\details{
\tabular{ll}{
Provider \tab Web \cr
GBIF \tab \url{http://www.gbif.org/} \cr
BISON \tab \url{http://bison.usgs.ornl.gov/} \cr
eBird \tab \url{http://ebird.org/content/ebird/} \cr
iNaturalist \tab \url{http://www.inaturalist.org/} \cr
Berkeley ecoengine \tab \url{https://ecoengine.berkeley.edu/} \cr
AntWeb \tab \url{http://www.antweb.org/} \cr
VertNet \tab \url{http://vertnet.org/} \cr
iDigBio \tab \url{https://www.idigbio.org/}
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Karthik Ram \email{karthik.ram@gmail.com}

Ted Hart \email{edmund.m.hart@gmail.com}
}
\keyword{package}

