% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{spocc_capwords}
\alias{spocc_capwords}
\title{Capitalize the first letter of a character string.}
\usage{
spocc_capwords(s, strict = FALSE, onlyfirst = FALSE)
}
\arguments{
\item{s}{A character string}

\item{strict}{Should the algorithm be strict about capitalizing. Defaults to FALSE.}

\item{onlyfirst}{Capitalize only first word, lowercase all others. Useful for
taxonomic names.}
}
\description{
Capitalize the first letter of a character string.
}
\examples{
 \dontrun{
spocc_capwords(c('using AIC for model selection'))
spocc_capwords(c('using AIC for model selection'), strict=TRUE)
}
}
\keyword{internal}

