% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spocc_rcharts_togeojson}
\alias{spocc_rcharts_togeojson}
\title{Get palette actual name from longer names}
\usage{
spocc_rcharts_togeojson(list_, lat = "latitude", lon = "longitude")
}
\arguments{
\item{list_}{A list}

\item{lat}{Latitude name}

\item{lon}{Longitude name}
}
\description{
Get palette actual name from longer names
}
\keyword{internal}

