% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wkt.R
\name{wkt}
\alias{wkt}
\alias{wkt_read}
\alias{wkt_write}
\title{Read or write WKT}
\usage{
wkt_read(wkt)

wkt_write(geojson)
}
\arguments{
\item{wkt}{(character) A Well Known Text string}

\item{geojson}{(character) GeoJSON as a list, character string or JSON}
}
\description{
Read or write WKT
}
\examples{
wkt <- 'LINESTRING (30 10, 10 30, 40 40)'
wkt_read(wkt)

wkt <- "POLYGON((38.4 -125,40.9 -125,40.9 -121.8,38.4 -121.8,38.4 -125))"
wkt_read(wkt)
}

